% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTSSR.R
\name{scTSSR}
\alias{scTSSR}
\title{use scTSSR to impute dropout values in scRNA-seq data}
\usage{
scTSSR(X_count, lambda1 = NULL, lambda2 = 1e+10, initA = NULL,
  initB = NULL, percent = 0.05, ncores = 1, MAX_ITER = 4,
  ABSTOL = 0.001, learning_rate = 1e-04, epochs = 100,
  batch_size = 128, run_batch = TRUE, verbose = TRUE,
  estimates.only = FALSE)
}
\arguments{
\item{X_count}{An expression count matrix. The rows correspond to genes and
the columns correspond to cells. Can be sparse.}

\item{lambda1}{Tuning parameter to facilitate feature selection and regularization.}

\item{lambda2}{Tuning parameter to penalize the diagonal elements of the parameter to
eliminate the trivial solution of representing an expression level as a linear combination
of itself.}

\item{initA}{The initionlization of A. The elements of A represent the similarities
between genes.}

\item{initB}{The initionlization of B. The elements of B represent the similarities
between cells.}

\item{percent}{The expression count matrix is preprocessed by filtering out the genes
expressed in at most percent*\eqn{100\%} of the cells.}

\item{ncores}{Number of cores to use. Default is 1.}

\item{MAX_ITER}{Maximum iteration of the external circulation of scTSSR.}

\item{ABSTOL}{Absolute tolerance of the external circulation.}

\item{learning_rate}{A hyper-parameter that controls the speed of adjusting the weights of the network
with respect to the loss gradient.}

\item{epochs}{The number of the entire training set going through the entire network.}

\item{batch_size}{The number of examples that are fed to the algorithm at a time.}

\item{run_batch}{Whether to use batch or to set the number of all the samples as the value of the batch size.
Default is TRUE.}

\item{verbose}{Whether to output the value of metrics at the end of each epoch. Default is TRUE.}
}
\value{
If `estimates.only = TRUE', then a matrix of scTSSR estimates.

If `estimates.only = FALSE', a list with the following components

\item{\code{estimate}}{Recovered (normalized) expression.}

\item{\code{se}}{Standard error of estimates.}

\item{\code{info}}{Information about dataset.}


The \code{info} element is a list with the following components:

\item{\code{size.factor}}{Size factor used for normalization.}

\item{\code{pred.time}}{Time taken to generate predictions.}

\item{\code{posterior.time}}{Time taken to compute the posterior distribution.}

\item{\code{total.time}}{Total time for scTSSR estimation.}
}
\description{
use scTSSR to impute dropout values in scRNA-seq data
}
\examples{

data("baron")

baron_imputation_result = scTSSR(baron$count.samp)

}
\author{
Ke Jin, \email{kej13@mails.ccnu.edu.cn}
}
