% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTSSR.R
\name{scTSSR}
\alias{scTSSR}
\title{use scTSSR to impute dropout values in scRNA-seq data}
\usage{
scTSSR(X_count, lambda = NULL, initA = NULL, initB = NULL,
  percent = 0.1, ncores = 1, MAX_ITER = 4, ABSTOL = 0.001,
  gamma = 1, beta = 0.5, max_iter = 100, abstol = 1e-04,
  penalize_diagonal = FALSE, estimates.only = FALSE)
}
\arguments{
\item{X_count}{An expression count matrix. The rows correspond to genes and
the columns correspond to cells. Can be sparse.}

\item{lambda}{Tuning parameter to facilitate feature selection and regularization.}

\item{initA}{The initionlization of A. The elements of A represent the similarities
between genes.}

\item{initB}{The initionlization of B. The elements of B represent the similarities
between cells.}

\item{percent}{The expression count matrix is preprocessed by filtering out the genes
expressed in at most percent*\eqn{100\%} of the cells.}

\item{ncores}{Number of cores to use. Default is 1.}

\item{MAX_ITER}{Maximum iteration of the external circulation of scTSSR.}

\item{ABSTOL}{Absolute tolerance of the external circulation.}

\item{gamma}{The step size.}

\item{beta}{The line search parameter.}

\item{max_iter}{Maximum iteration of the accelerated proximal gradient descent algorithm.}

\item{abstol}{Absolute tolerance of the accelerated proximal gradient descent algorithm.}

\item{penalize_diagonal}{Whether penalize the diagonal elements of
the regression coefficient matrix. Default is FALSE.}

\item{estimates.only}{Only return the imputed data. Default is FALSE.}
}
\value{
If `estimates.only = TRUE', then a matrix of scTSSR estimates.

If `estimates.only = FALSE', a list with the following components

\item{\code{estimate}}{Recovered (normalized) expression.}

\item{\code{se}}{Standard error of estimates.}

\item{\code{info}}{Information about dataset.}


The \code{info} element is a list with the following components:

\item{\code{size.factor}}{Size factor used for normalization.}

\item{\code{pred.time}}{Time taken to generate predictions.}

\item{\code{posterior.time}}{Time taken to compute the posterior distribution.}

\item{\code{total.time}}{Total time for scTSSR estimation.}
}
\description{
use scTSSR to impute dropout values in scRNA-seq data
}
\examples{

data("baron")
baron_imputation_result = scTSSR(baron$count.samp)

}
\author{
Ke Jin, \email{kej13@mails.ccnu.edu.cn}
}
